+++ src/presentation/controllers/singup/singup-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../protocols'
 export * from '../../../domain/useCases/add-account'
 export * from '../../../domain/models/account'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
