+++ src/presentation/helpers/validators/compare-fields-validation.ts
@@ -1,5 +1,5 @@
 import { InvalidParamError } from '../../errors'
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 
 export class CompareFieldValidation implements Validation {
   private readonly fieldName: string
